Add date started and completed to BookDetailScreen, Rename composable packages to components
