+++ app/src/main/java/com/enoch02/literarylinc/ui/theme/Theme.kt
@@ -2,6 +2,8 @@ package com.enoch02.literarylinc.ui.theme
 
 import android.app.Activity
 import android.os.Build
+import androidx.compose.animation.AnimatedContent
+import androidx.compose.animation.ExperimentalAnimationApi
 import androidx.compose.foundation.isSystemInDarkTheme
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.darkColorScheme
